FIX tseries can't pull yahoo data, sfu mirror gone
Started from report of an R problem in Webwork. The problem tries to
pull historical financial data from yahoo using the tseries library and
fails to connect. Looks like tseries started depending on the quantmod
library for pulling yahoo data, so we needed to update tseries to
latest.

Unfortunately, the Dockerfile wouldn't build anymore, as the SFU debian
mirrors appear to have disappeared. So I've bumped up the debian version
to bullseye. This resulted in R being bumped from 3.6.1 to 4.0.4-1. From
the R release notes, I don't see any breaking changes, so hopefully it
should be a seamless upgrade.

Some of the R libraries failed to install from cran, so I've installed
the debian repo of them instead.

The Rserve available in the Debian repo is 1.7.3 which is older than the
1.8.6 we are using, so I've just retained 1.8.6 for now.