FIX github action failed to push to docker hub
When trying to push, failed with 'requested access to the resource is
denied'. Found similar issues others had and they fixed it by upgrading
docker/build-push-action. Since this github action was copied from
ComPAIR, it was on docker/build-push-action@v1. Not sure how it still
works in ComPAIR, maybe it was grandfathered in somehow? Apparently,
we're supposed to have a separate docker/login-action stage that does
the login to docker.

Updating to docker/build-push-action@v4 means the tag_with_ref option is
no longer supported, so had to add a docker/metadata-action stage to
grab the git tags for use as docker tags.

Correction, turns out I was using the wrong secret variables. So the
upgrade probably wasn't necessary, but keeping it anyways.